Prucimg shalu stable

+++ automation/src/main/java/ocean/test/condition/pricing/OCEAN_Pricing_TC_23_24.java
@@ -1,5 +1,6 @@
 package ocean.test.condition.pricing;
 
+import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNotEquals;
 import static org.testng.Assert.assertTrue;
 
@@ -7,7 +8,6 @@ import java.util.HashMap;
 
 import org.testng.annotations.Test;
 
-import ocean.modules.dataprovider.PricingDataProvider;
 import ocean.modules.pages.PricingModulePages;
 
 /**
@@ -24,48 +24,47 @@ public class OCEAN_Pricing_TC_23_24 extends PricingModulePages {
 	 * Validate classification list for cloning/exporting and delete.this function
 	 * cover the condition of this test case like 1. Cloning ClassificationList
 	 */
-	@Test(priority = 1, groups = "regression", dataProvider = "fetchDataForTC24", dataProviderClass = PricingDataProvider.class, description = "Validate classification list for cloning.")
-	public void validateClassificationListForCloneExport(String[] inputArray) throws Exception {
+	@Test(priority = 5, groups = "regression", description = "Validate classification list for cloning.")
+	public void validateClassificationListForCloneExport() throws Exception {
 		goToPricingTab();
 		goToClassficationList();
-		HashMap<String, String> map = new HashMap<String, String>();
-		map.put("PriceSheetName", inputArray[0]);
-		selectCloneClassfication(map);
-		validateClonePriceSheet(map);
+		selectCloneClassfication("ClonePriceSheet");
+		validateClonePriceSheet("ClonePriceSheet");
 		String validationMessage = getValueofClonePriceSheet();
-		assertTrue(validationMessage.equalsIgnoreCase(map.get("PriceSheetName")), "Not matched");
+		assertTrue(validationMessage.equalsIgnoreCase("ClonePriceSheet"));
 	}
 
 	/**
 	 * Validate classification list for cloning/exporting and delete.this function
 	 * cover the condition of this test case like 2. Export ClassificationList
 	 */
-	// @Test(priority = 2, groups = "regression", description = "Validate
-	// classification list for exporting.")
+	@Test(priority = 5, groups = "regression", description = "Validate classification list for exporting.")
 	public void validateClassificationListForExport() throws Exception {
 		goToPricingTab();
 		goToClassficationList();
 		exportClassificationPriceSheet();
-		String downloadPath = "C:\\Users\\shalu.chauhan\\Desktop";
+		String desktopPath = System.getProperty("user.home") + "\\Desktop";
 		String fileName = getTextOfElement("getExportClassificationListName");
-		isFileDownloaded(downloadPath, fileName);
-		String deletepath = "C:\\Users\\shalu.chauhan\\Desktop\\Absolute Lease - ACL_Master_Export.xlxs";
-		deleteExportFile(deletepath);
+		boolean deleted = isFileDownloaded(desktopPath, fileName);
+		if (deleted == true) {
+			String deletepath = desktopPath + "\\" + fileName;
+			deleteExportFile(deletepath);
+			boolean deleted1 = isFileDownloaded(desktopPath, fileName);
+			assertEquals(deleted1, false);
+		} else {
+			throw new Exception("File not deleted");
+		}
 	}
 
 	/**
 	 * Validate classification list for cloning/exporting and delete.this function
 	 * cover the condition of this test case like 3. Delete ClassificationList
 	 */
-	// @Test(priority = 3, groups = "regression", dataProvider = "fetchDataForTC24",
-	// dataProviderClass = PricingDataProvider.class, description = "Validate
-	// classification list for delete.")
-	public void validateClassificationListForDelete(String[] inputArray) throws Exception {
+	@Test(priority = 5, groups = "regression", description = "Validate classification list for delete.")
+	public void validateClassificationListForDelete() throws Exception {
 		goToPricingTab();
 		goToClassficationList();
-		HashMap<String, String> map = new HashMap<String, String>();
-		map.put("PriceSheetName", inputArray[0]);
-		deleteClassificationPriceSheet(map.get("PriceSheetName"));
+		deleteClassificationPriceSheet("ClonePriceSheet");
 		validateDeleteClassificationPriceSheet("listOfClassificationList");
 	}
 
@@ -75,22 +74,17 @@ public class OCEAN_Pricing_TC_23_24 extends PricingModulePages {
 	 * classification list when it is increased by one on a vehicle for turbo
 	 * option.
 	 */
-	// @Test(priority = 4, groups = "regression", dataProvider = "fetchDataForTC24",
-	// dataProviderClass = PricingDataProvider.class, description = "Validate usage
-	// of classification list when it is increased by one on a vehicle for turbo
-	// option.")
-	public void validateClassificationListForTurbo(String[] inputArray) throws Exception {
+	@Test(priority = 5, groups = "regression", description = "Validate usage of classification list when it is increased by one on a vehicle for turbo option.")
+	public void validateClassificationListForTurbo() throws Exception {
 		goToPricingTab();
 		goToClassficationList();
-		HashMap<String, String> map = new HashMap<String, String>();
-		map.put("Turbo", inputArray[2]);
-		map.put("ClassificationListForTurbo", inputArray[1]);
-		selectPricesheetForTurbo(map);
-		HashMap<String, String> actualTurboValue = getClassficationListForTurboBeforAndAfterEditing(inputArray[1], 0);
-		editClassificationListForTurbo(map);
-		HashMap<String, String> expectedTurboValue = getClassficationListForTurboBeforAndAfterEditing(inputArray[1], 1);
-		uncheckedEditTurboCheckbox();
-		waitForSomeTime(3);
+		HashMap<String, String> actualTurboValue = getClassificationInfomation("", "0", "0");
+		selectPricesheetForTurbo(actualTurboValue.get("CLASS_GROUP_NAME"));
+		editClassificationListForTurboAndDiesel("Turbo", "Y");
+		checkEnableDisable("clickCheckBoxOfTurbo");
+		HashMap<String, String> expectedTurboValue = getClassificationInfomation(
+				actualTurboValue.get("CLASS_GROUP_NAME"), "1", "0");
+		editClassificationListForTurboAndDiesel("Turbo", "N");
 		assertNotEquals(actualTurboValue, expectedTurboValue);
 	}
 
@@ -100,24 +94,16 @@ public class OCEAN_Pricing_TC_23_24 extends PricingModulePages {
 	 * classification list when it is increased by one on a vehicle for turbo
 	 * option.
 	 */
-	// @Test(priority = 5, groups = "regression", dataProvider = "fetchDataForTC24",
-	// dataProviderClass = PricingDataProvider.class, description = "Validate usage
-	// of classification list when it is increased by one on a vehicle for turbo
-	// option.")
-	public void validateClassificationListForDiesel(String[] inputArray) throws Exception {
+	@Test(priority = 5, groups = "regression", description = "Validate usage of classification list when it is increased by one on a vehicle for turbo option.")
+	public void validateClassificationListForDiesel() throws Exception {
 		goToPricingTab();
 		goToClassficationList();
-		HashMap<String, String> map1 = new HashMap<String, String>();
-		map1.put("Diesel", inputArray[3]);
-		map1.put("ClassificationListForDiesel", inputArray[4]);
-		selectPricesheetForDiesel(map1);
-		HashMap<String, String> actualDieselValue = getClassficationListForDieselBeforeAndAfterEditing(inputArray[4],
-				0);
-		editClassificationListForDiesel(map1);
-		HashMap<String, String> expectedDieselValue = getClassficationListForDieselBeforeAndAfterEditing(inputArray[4],
-				1);
-		uncheckedEditDieselCheckbox();
-		waitForSomeTime(3);
+		selectPricesheetForDiesel("ACL");
+		HashMap<String, String> actualDieselValue = getClassficationListForDieselBeforeAndAfterEditing("ACL", 0);
+		editClassificationListForTurboAndDiesel("Diesel", "Y");
+		checkEnableDisable("clickCheckBoxOfDiesel");
+		HashMap<String, String> expectedDieselValue = getClassficationListForDieselBeforeAndAfterEditing("ACL", 1);
+		editClassificationListForTurboAndDiesel("Diesel", "N");
 		assertNotEquals(actualDieselValue, expectedDieselValue);
 	}
 }
